feat: implement silk wallet
This is done as a connector for Wagmi, so the code is nicely contained
and clean. That means, however, that we're losing browser wallet support
for the moment.

Moved the wagmi config into wagmi/config.ts.

FYI: There's no update to bun.lockb since I've cherry-picked the
tracking of this file (70eb49ce8ff7d646b55549bba5d8bd4bf9214d7a) from an
earlier version of this branch that had silk already installed.